+++ acceptance/tests/modules/build/build_should_not_create_changes.rb
@@ -9,6 +9,12 @@ defaultversion = '0.1.0'
 buildpath = "#{modname}/pkg/#{modauthor}-#{modname}-#{defaultversion}"
 
 agents.each do |agent|
+
+  if on(agent, facter("fips_enabled")).stdout =~ /true/
+    puts "Module build, loading and installing not supported on fips enabled platforms"
+    next
+  end
+
   teardown do
     on(agent, "rm -rf #{modname}")
   end
